requirements.txt
flask==3.0.2
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.1.0
imutils==0.5.4
pdf2image==1.16.3  # For PDF support
python-Levenshtein==0.25.1  # Fuzzy text matching